optimize helm deploy logic
